made some import changes to transmission line functions. i need to fix this conceptually.
should electrical length return a negative number???
should a WorkingBand.line be defined as exp(-j*theta) or +,
should Gamma0_at_theta be defined in a similar maner? 

the answer to these questions is dependent on how i defined the sign of propagation 
constant.  but, should electrical length be signed???? is is this needed?