changed the order which touchstone loads a file, so it raises a io error if file doesnt exist as opposed to a touchstone extension error.
moved from argparser to optparser in plot_touchstone, which is 
a bummer but i had to do it for compatability with older pythons.

i also changed the name loaded into a network type to be 
just the filename with no absolute path and no extension using 
standard os.path calls.