removed `frequency` argument of Calibration initializer.
the frequency argument is redundant, because each Network
provides a Frequency object containing the relevent
frequency information.

also i made the input measured and ideals lists, copied,
element-wise, so that the Calibration object can alter the
Networks themselves.  this nice because the calibration
object can try to interpolate if the frequency info
doesnt match up.

also and updated testCase to not use the argument either.