fixed generate_subnetworks correctly
this will be called after any primary property is set, so either
s,z, or y. the names are still s11, s12, etc, which is maybe
misleading as they are not dependent upon the s-parameter
representation. however, it seems sensible to me, so untill i
think of something better ill leave it.