fix impedance step bug, shelled out s-param tests, changed watermark
fixed a major conecptual bug in teh way an impedance step is
calculated. the root of the problem being that used transmission
coefficient instead of the scattering parameter definition.

this leads to a more important point; skrf's treatment of s-parameters
is over-simplified. looking at
 Kurokawa, K., "Power Waves and the Scattering Matrix"

as well as teh qucs/hfss docs i see that my implmemntation is only
valid for a special case. when i get around to it, i will implement
this correctly which will allow for s-parmeter re-normalization.