fixing Frequency, and Network to accept arbitrary frequency vectors
modified `Frequency` class's `from_f()` function, and changed
`Network.read_touchstone()` to use this.

created  a test_frequency test case to test the from_f and the
linear sweep init.

it would be nice to change the initializer to accept both a
start/stop/n_points or an vector, with  unit. instead of doing
the class method, but will have to think about it