fixing Touchstone class to take file objects
fixed elusive bug this created in NetworkSet.from_zip()

i want to allow the read functions to accept a filename, or a
file-object. also, to use the simpler try/except blocks to read
a file (which is nice because it doesnt rely on any extension
testing) the file object must be reset, if it is left open.

so a convention must be adopted:
if a file-object is passed to a function
   leave file open, seek to begginning if obj creation fails
if a filename is passed,
   close file at the the end of the function