port impedance variable rename
in order to be backwards compatable, i re-name port_z0 -> z0, as
it was, and renamed the characteristic_impedance z0->Z0.

this is little confusing, but its nice because the port impedance
of Media is the same as the Network.

the Media's characteristic impedance is actually less used, and
could be renamed in the future, z0_char or eta0 ,