remove meas. parameters declarations and functions
This commit fixes/addresses issue #178. The assignment of default
measurement settings is not the scope of the VNA superclass. Removing
these declarations and functions will probably break some
implementations (see #178 for details).

Furthermore, the VNA superclass should not have a default address
assigned and should always be called with a specific address. If a
default address should be given in the subclasses is another discussion.