Transpose 2-port data
In TouchStone 1.0 files, the data order for 2-port data is:
(11) (21) (12) (22)

while for a higher number of ports, the data is arranged as follows
(11) (12) (13) ... (1n)

This is now honored in the get_sparameter_data() function's return
values, if the filename ends on "s2p". The internal representation
of the class is unchanged.