Admittance Smith chart labels fixed
rf.plotting.smith(chart_type='y', draw_labels=True)
one sees that the labels for the real and imag part of the admittances are incorrect. The real part should go (from left to right from /infty to 0.0, and the imag part on the unit circle goes (clock wise) from \infty j over -j to 0, +j and back to \infty.