Reading new gamma/z0 HFSS format
Since HFSS 2020R2, the gamma and z0 are written on a single line instead of many in the previous version. Add tests with multiport files
added pathlib to the dependencies list for python 2

debug messages for Travis


put decimals to integer before division


removing the print in the test. Should be almost final now