add use_z_instead_ifft
in the IEEEP repo, IEEEP370 NZC has the ability to use `s2z` and `z2s` instead ifft.

According to the commit that added this functionnality, it is intended to use with very short fixture that have not enough point in time domain,  e.g. a single point. In this case the impedance midpoint computation cannot give the expected results.