Three Receivers Calibration: clarify SOLT procedures.
This update clarifies the Correct SOLT Calibration Procedure
for 1.5-port VNAs. First, I added a reference to the SOLT example.
I also explained the difference between a standard SOLT calibration
and a three-receiver calibration, including the fact that SOL at
port 2 is not needed (unless there's need to calibrate isolation),
and that the results are still saved as 2-port network, but S22 and
S12 will contain no meaningful data.

This clarifies the question reported in Issue #629.

Signed-off-by: Yifeng Li <tomli@tomli.me>