Update general.py
The reading of files is not traceable or inconsistent, especially for linux machines. 
Filenames were chosen arbitrary and a simple sorted() could fix it.

I already submitted somehow same request in 2020.

I would even recommend to set sorted even as standard. I do not know why this request has been removed again afterwards, probably just by error or there is a specific reason. Probably there's a reason for the backchange I'm not aware of or missed when skimming the code for the first time.

I did not set it the sorted as standard so far because of above mentioned reasons. 

Please consider this change, as I am really struggling with it. 
Thanks in advance.