SOLT.ipynb: distinguish offset and port impedances, close #754.
Currently, in example code snippets, if a specific offset impedance
is used, the VNA port impedance is also set to this matching value.
The original justification was that, this way, specifying a 75-ohm
offset impedance will create S-parameters as measured by a 75-ohm VNA,
which is a reasonable default.

However, sometimes an offset impedance value slightly different from
the port impedance is used to model imperfections in the standard.
For example, a 50-ohm VNA standard may have a 50.209-ohm or 49.992-ohm
offset impedance. Setting port impedance to offset impedance in this
case is misleading and incorrect.

This commit updates the guide to highlight this fact in the explanation.
The example code snippets are also updated with an additional parameter,
allowing users to explicitly specify different offset and port impedances,
followed by 3 usage examples.

Also, a brief explanation of the deprecation warning is added.

Signed-off-by: Yifeng Li <tomli@tomli.me>