Remove `html_theme_path` setting from docs config to fix search/JS
Currently search and other JavaScript functions fail on the RTD
documentation page, due to an exception thrown over missing jQuery.

Searching upstream, this can be caused by setting `html_theme_path`
which was previously recommended but is now out of date.

This updates the config in-line with current advice from sphinx_rtd_theme
adding it to `extensions` and removing the `html_theme_path` setting.