doc/SOLT Calibration Standards Creation: rewrite, close #1308.
Completely rewrite the article "SOLT Calibration Standards Creation".
The new article includes many changes:

1. Incorrect impedance renormalization #1308 has been fixed.

2. A comprehensive review of the physical meaning of calibration standard
parameters and their practical implications, rather than just providing the
means of calculation.

3. An alternative formulation is used to define calibration standard in
terms of an RLCG equivalent circuit, this is more physically meaningful than
the raw propagation constants. The disadvantage is that the method doesn't
reproduce the exact Keysight's official numerical values (because Keysight
uses a low-loss approximation). Given the fact that Keysight 85056A's
magnitude and phase angle are in agreement to 4 decimal places at 50
GHz, this difference is only a theoretical problem.

4. In case that numerical consistency is important, the derivation and
implementation of Keysight's official formulas have also been provided
in an Appendix.

5. The difference of "flush" (zero-length) and "offset" standards is
explained, including a photograph. An additional practical example of
generic SMA calibration kit (commonly seen in amateur radio applications)
is given as an example of "flush" standard, with personal measurement
results.

6. Note that a signal travels in both directions in a S11 measurement,
so the loss from S11 in decibel is doubled numerically in comparison to
the stated offset loss.

7. Years after the original tutorial has been written, 4 out of 4 links
in it are dead. This new article replaces all links with archive.org or
DOI links to ensure future accessibility. All citation entries are also
in an IEEE-like format with proper attributions to their sources, so
hopefully people can find them independently even if archive.org is down.

Signed-off-by: Yifeng Li <tomli@tomli.me>