media.py: reword misleading z0 documentation, close #1310.
The documentation of load(), match(), short(), and thru() says:

> z0 (number, or array-like or None) – port impedance. Default is None,
> in which case the Media’s z0 is used. This sets the resultant Network’s z0.

But the documentation is misleading. Clearly, the wording was not updated
after the new impedance API syntax has landed. This is problem is important
because the documentation for capacitor() and inductor() redirects readers
to match(). Without clarification, the reference impedance of inductance
or capacitance will be confusing.

The new wording is:

> z0 (number, or array-like or None) – port impedance.
> This sets the resultant Network’s z0. Default is None, the Media’s
> z0_port value is used. If z0_port is also None (by default), in which case
> the Media's native z0. A Media's z0_port is an optional attribute for
> renormalizing resultant Network to a new reference impedance that differs
> from a Media's native z0.

Signed-off-by: Yifeng Li <tomli@tomli.me>