calibration.py: fix incorrect UnknownThru calibration as reported in #1317.
In skrf.calibration.calibration.UnknownThru, an undetected branch cut bug
exists: occasionally, UnknownThru chooses the incorrect branch of the solved
square root for the error term "k", causing invalid calibration with
180-degree phase jump.

This bug only occurs when the Thru is lossy with significant attenuation
(e.g. 5 dB), and when many data points are tested, thus this bug has never
been detected by the test suite. Originally, the bug was detected using a
waveguide medium renormalized to 50 Ω, thus unintentionally triggering the
bug because of the loss introduced via the renormalization process. Later,
it has been confirmed that coaxial mediums are affected as well by introducing
losses in the Thru.

Having prior knowledge of the square root sign of the term k is a known
limitation of the UnknownThru algorithm itself, but the current estimation
method is not unreliable. Currently, the sign is estimated by comparison
with a standard EightTerm calibration using the user-supplied Thru
approximation. This is not reliable if the actual Thru is not so ideal
(e.g. with losses).

To workaround the problem, we simply run the UnknownThru calibration twice
with both solutions. The solved Thru network is compared against the
user-supplied approximate definition, so that the correct term k can be
determined at each frequency point.

Signed-off-by: Yifeng Li <tomli@tomli.me>