calibration.py: fix incorrect MRC calibration, close #1317.
The MRC calibration is a combination of SDDL one-port calibration and
UnknownThru two-port calibration. Being a derivative of the UnknownThru,
it's affected by the same calibration bug #1317.

We noticed that the MRC class is nearly a copy-paste of the UnknownThru
calibration, so it's not necessarily to duplicate the fix in UnknownThru
again. Instead, we introduce a class constant called _ONEPORT_ALGO,
which is set to OnePort in UnknownThru, but set to SDDL in MRC. This way
only a single constant definition needs to be changed.

Signed-off-by: Yifeng Li <tomli@tomli.me>