calibration.py: implement non-50 ohm calibration, close #902 #1307 #1313
This commit implements calibration for non-50 ohm system impedance by
correctly tracking the reference impedance of the DUT, the error networks,
and the calibration coefficients. The logic works as follows:

1. Ensure all calibration standard definitions and measurements have
the same reference impedance at all ports, and for all frequencies.

2. Add a new entry to the 3 terms, 8 terms and 12 terms calibration
coefficients, named "system impedance". The system impedance is
extracted from the reference impedance of the first measured network
(this information extraction convention is already used in the Calibration
class).

3. When running apply_cal(), each calibration subclass must write this
information to the coefficient dictionary. When creating error networks,
the system impedance is used as the reference impedance. Since the reference
impedance is always imported and exported together with the dictionary,
a reconstructed calibration instance preserves the information.

4. It's named the "system impedance" to highlight the fact that its the
reference impedance used to calculate the error terms and error boxes,
its behavior is similar to the VNA measurement systems: a global
setting, constant at all ports, and is unchanged even after TRL calibration.

5. Mixed port reference impedance is not implemented. Currently, all
calibration algorithms assume a constant system impedance. This feature
will be extremely difficult, if ever possible to support, a major redesign
would be necessary. However, implementing per-frequency reference impedance
is in principle possible and much easier. Thus, for future extensibility,
the "system impedance" is a 1D array at each calibrated frequency point.

Signed-off-by: Yifeng Li <tomli@tomli.me>