mathFunctions.py: API to set/get the global numpy RNG, close #1314.
This commit introduces rand_set_rng() and rand_rng() for
setting and getting the global numpy RNG used by scikit-rf.
This is useful for fixing a random seed for reproducible
Monte Carlo analysis and unit tests.

This function is expected to be called before using any
scikit-rf features, since it's a global variable and thread-unsafe.
To temporarily change the random number generator of a particular
method (e.g. skrf.media.random), use the `rng` argument instead.

Signed-off-by: Yifeng Li <tomli@tomli.me>