calibration.py: raise ValueError if 2-port inputs are used, close #1321.
In the current calibration code, OnePort class checks whether users
accidentally use 2-port input networks for 1-port calibration, and
raises a RuntimeError if so. But there's no check for SDDL, SDDLWeikle,
and PHN calibrations. As a result, they all threw different errors in
the middle of a calibration run, sometimes it's IndexError, sometimes
it's ValueError.

This commit adds a class-wide self._check_input() methods for all one-port
calibration classes, which is called in self.run(). For consistency, we
use ValueError instead of the original RuntimeError.

Signed-off-by: Yifeng Li <tomli@tomli.me>